requirements.txt
amqp==1.4.9
anyjson==0.3.3
attrs==21.4.0
backports.functools-lru-cache==1.6.4
billiard==3.3.0.23
celery==3.1.24
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
click==6.7
configparser==4.0.2
contextlib2==0.6.0.post1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.2
Django==1.11.20
django-activity-stream==0.6.5
django-autoadmin==1.1.1
django-autocomplete-light==3.2.10
django-bootstrap-modal-forms==2.2.0
django-celery==3.2.1
django-colorful==1.2
django-cors-headers==2.2.0
django-downloadview==1.9
django-environ==0.4.5
django-formtools==2.1
django-js-asset==1.2.3
django-mathfilters==0.4.0
django-model-utils==3.1.2
django-mptt==0.9.1
django-pure-pagination==0.3.0
django-pylibmc==0.6.1
django-qsstats-magic==1.0.0
django-solo==1.1.3
django-stronghold==0.3.0
django-suit==0.2.26
django-widget-tweaks==1.4.2
djangorestframework==3.7.7
djangorestframework-recursive==0.1.2
drf-yasg==1.5.0
enum34==1.1.10
flanker==0.9.0
flex==6.13.2
fpdf==1.7.2
funcsigs==1.0.2
functools32==3.2.3.post2
furl==1.2
fusepy==2.0.4
future==0.18.2
gevent==1.3.5
graphviz==0.8.4
greenlet==0.4.15
gunicorn==19.9.0
idna==2.6
importlib-metadata==2.1.3
inflection==0.3.1
ipaddress==1.0.23
itypes==1.2.0
Jinja2==2.11.3
jsonpointer==1.14
jsonschema==3.2.0
kombu==3.0.37
MarkupSafe==1.1.1
mock==2.0.0
node-semver==0.3.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
pathlib2==2.3.2
pbr==5.11.0
Pillow==5.2.0
ply==3.11
psycopg2==2.7.3.2
pycountry==18.5.26
pycparser==2.21
pylibmc==1.6.1
pyocr==0.5.2
PyPDF2==1.26.0
pyrsistent==0.16.1
python-dateutil==2.7.3
python-gnupg==0.3.9
python-magic==0.4.15
pytz==2018.3
PyYAML==3.13
redis==2.10.6
regex==2019.4.14
requests==2.18.4
rfc3987==1.3.8
ruamel.ordereddict==0.4.15
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
scandir==1.10.0
sh==1.12.14
six==1.16.0
strict-rfc3339==0.7
swagger-spec-validator==2.1.0
tld==0.12.6
typing==3.10.0.0
uritemplate==3.0.1
urllib3==1.22
validate-email==1.3
WebOb==1.8.7
whitenoise==3.3.1
zipp==1.2.0
